Move gzifstream and gzofstream to io namespace.
They make more sense there, not sure why I put them in util...
