+++ src/io/tools/CMakeLists.txt
@@ -1,4 +1,4 @@
 if (ZLIB_FOUND)
     add_executable(compressor-test compressor_test.cpp)
-    target_link_libraries(compressor-test meta-util)
+    target_link_libraries(compressor-test meta-io)
 endif()
